requirements.txt
edge_tts==6.1.10
google-generativeai==0.5.3
keyboard==0.13.5
obs-websocket-py==1.0
openai==1.13.3
openai-whisper==20231117
protobuf==4.25.3
pyaudio==0.2.14
pydub==0.25.1
pyside6==6.6.2
pytchat==0.5.5
pyvts==0.3.2
Requests==2.31.0
tiktoken==0.7.0
transformers==4.38.2
twitchio==2.9.1